Add a RestTest function to test utils
- add models folder
- add TestRun model holding information of one test and all the calls
from that test
- add TestCase model holding information about a test case
- add RestMethod Enum listing all methods used by Menas
- add TestCaseResult to model holding results of previously mentioned
TestCases
- add CmdConfig for a spark job running the rest tests
- add RestCaller Object which is taking care of choosing the correct
methods, comparing outputs and meassuring time it took for the test case
to finish.
- add RestRunnerJob to run all test in the same way enceladus projects
are acustomed to. Load data, runs all the rest calls and save evaluation
to a file. Added to this output is a _TEST_SUMMARY file which saves the
test name, start and end time, number of failed and passed test cases
and other information.