#256 Replace maven enforcer with antrun copy task
Let's replace the enforcer plugin, which fails the build if the
application.conf is missing from the main resources.

Rather let's materialize the template on the first build, which
will provide reasonable defaults without necessary providing
custom configs while preserving existing files.

This means that people should be able to clone and build the
project (incl tests) out of the box.

Signed-off-by: Jan Scherbaum <kmoj02@gmail.com>